int -> long changes update input types for cnns to properly handle different weight formats as well as support different rnn formats spacing changes

+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/params/DeconvolutionParamInitializer.java
@@ -49,8 +49,8 @@ public class DeconvolutionParamInitializer extends ConvolutionParamInitializer {
         org.deeplearning4j.nn.conf.layers.Deconvolution2D layerConf =
                 (org.deeplearning4j.nn.conf.layers.Deconvolution2D) conf.getLayer();
         if (initializeParams) {
-            int[] kernel = layerConf.getKernelSize();
-            int[] stride = layerConf.getStride();
+            long[] kernel = layerConf.getKernelSize();
+            long[] stride = layerConf.getStride();
 
             val inputDepth = layerConf.getNIn();
             val outputDepth = layerConf.getNOut();
@@ -65,7 +65,7 @@ public class DeconvolutionParamInitializer extends ConvolutionParamInitializer {
 
             return weights;
         } else {
-            int[] kernel = layerConf.getKernelSize();
+            long[] kernel = layerConf.getKernelSize();
 
             INDArray weights =  WeightInitUtil.reshapeWeights(
                     new long[] {layerConf.getNIn(), layerConf.getNOut(), kernel[0],
@@ -81,7 +81,7 @@ public class DeconvolutionParamInitializer extends ConvolutionParamInitializer {
         org.deeplearning4j.nn.conf.layers.Deconvolution2D layerConf =
                 (org.deeplearning4j.nn.conf.layers.Deconvolution2D) conf.getLayer();
 
-        int[] kernel = layerConf.getKernelSize();
+        long[] kernel = layerConf.getKernelSize();
         val nIn = layerConf.getNIn();
         val nOut = layerConf.getNOut();
